+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v1.3.7.0 [unreleased]
 ### Bug Fixes
 1. [#1845](https://github.com/influxdata/chronograf/pull/1845): Fix no-scroll bar appearing in the Data Explorer table
+1. [#1864](https://github.com/influxdata/chronograf/pull/1864): Fix Write Data form upload button and add `onDragExit` handler
 
 ### Features